Add total liabilities / current assets ratio

+++ src/main/java/com/myproject/dto/yahoo/fundamental/v10/financial/data/FinancialDataDto.java
@@ -10,6 +10,6 @@ public class FinancialDataDto {
     private FinancialReportItem ebitda;
 
     public BigDecimal getEBITDA() {
-        return BigDecimal.valueOf(ebitda.getRawValue());
+        return ebitda.getRawValue();
     }
 }
